requirements.txt
hyperon==0.1.7
Flask == 3.0.3
Werkzeug==3.0.2
biocypher>=0.5.4
python-dotenv==1.0.1
PyYAML==6.0.1
flask-cors == 4.0.1
Flask-Mail ==0.10.0
openpyxl == 3.1.5
PyJWT == 2.9.0
pymongoose == 1.3.8
openai == 1.51.2
pytest == 8.3.3